requirements.txt
Flask==3.0.3
openai==0.28.0
python_docx==1.1.0
